renamed files and turned components into jsx instead of js
