+++ src/Components/footer.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Nav, Navbar, Row, Col } from "react-bootstrap";
 import { useLocation } from "react-router-dom";
-import BackToTop from "./BackToTop";
+import BackToTop from "./backtotop";
 
 //sets up props for the page
 const PageFooter = (props) => {
