+++ src/Components/homepage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Row, Col, Image } from "react-bootstrap";
 import { Link } from "react-router-dom";
-import Slideshow from "./Carousel";
+import Slideshow from "./carousel";
 
 // This is the homepage, including the carousel and below it, a row of recipe images and titles pulled from the BlogData component. //
 // The row of images resemble the one in RecipeIndex although vary in styling.//
